+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## 16.06.8
 - the menu entry `Show in system tray` was renamed to `Show menu bar item` on 
   OS X (for [Issue #214](https://github.com/pbek/QOwnNotes/issues/214))
+- the `Show toolbar` menu entry is now checked as long as there is one toolbar
+  activated (for [Issue #213](https://github.com/pbek/QOwnNotes/issues/213))
+    - if the menu entry it is checked by a user all toolbars will be activated 
+      again
 
 ## 16.06.7
 - when doing a text search in your note the selected text is now preset as